[#8597] Convert Sequence Data Type into Int

+++ profiler-test/src/test/java/com/navercorp/pinpoint/test/OrderedSpanRecorderTest.java
@@ -72,12 +72,12 @@ public class OrderedSpanRecorderTest {
         final TraceRoot traceRoot = new DefaultTraceRoot(traceId, agentId, startTime, 0);
 
         Span span = createSpan(traceRoot, startTime);
-        SpanChunk event = wrapSpanChunk(traceRoot, createSpanEvent(traceRoot, 0, (short) 0));
-        SpanChunk event1 = wrapSpanChunk(traceRoot, createSpanEvent(traceRoot, 0, (short) 1));
-        SpanChunk event2 = wrapSpanChunk(traceRoot, createSpanEvent(traceRoot, 0, (short) 2));
-        SpanChunk asyncEvent1_1 = wrapSpanChunk(traceRoot, createAsyncSpanEvent(traceRoot, 0, (short) 0), new DefaultLocalAsyncId(1, (short) 1));
-        SpanChunk asyncEvent1_2 = wrapSpanChunk(traceRoot, createAsyncSpanEvent(traceRoot, 0, (short) 1), new DefaultLocalAsyncId(1, (short) 1));
-        SpanChunk asyncEvent2 = wrapSpanChunk(traceRoot, createAsyncSpanEvent(traceRoot, 0, (short) 0), new DefaultLocalAsyncId(2, (short) 1));
+        SpanChunk event = wrapSpanChunk(traceRoot, createSpanEvent(traceRoot, 0, 0));
+        SpanChunk event1 = wrapSpanChunk(traceRoot, createSpanEvent(traceRoot, 0, 1));
+        SpanChunk event2 = wrapSpanChunk(traceRoot, createSpanEvent(traceRoot, 0, 2));
+        SpanChunk asyncEvent1_1 = wrapSpanChunk(traceRoot, createAsyncSpanEvent(traceRoot, 0, 0), new DefaultLocalAsyncId(1, (short) 1));
+        SpanChunk asyncEvent1_2 = wrapSpanChunk(traceRoot, createAsyncSpanEvent(traceRoot, 0, 1), new DefaultLocalAsyncId(1, (short) 1));
+        SpanChunk asyncEvent2 = wrapSpanChunk(traceRoot, createAsyncSpanEvent(traceRoot, 0, 0), new DefaultLocalAsyncId(2, (short) 1));
 
         final List<SpanType> expectedOrder = Arrays.asList(
                 span,
@@ -123,14 +123,14 @@ public class OrderedSpanRecorderTest {
 
 
         Span span = createSpan(traceRoot1, startTime1);
-        SpanChunk event1 = wrapSpanChunk(traceRoot1, createSpanEvent(traceRoot1, 0, (short) 0));
-        SpanChunk asyncEvent1_1_1 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 0, (short) 0), new DefaultLocalAsyncId(1, (short) 1));
-        SpanChunk asyncEvent1_1_2 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 0, (short) 1), new DefaultLocalAsyncId(1, (short) 1));
-        SpanChunk asyncEvent1_2_1 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 0, (short) 0), new DefaultLocalAsyncId(1, (short) 2));
+        SpanChunk event1 = wrapSpanChunk(traceRoot1, createSpanEvent(traceRoot1, 0, 0));
+        SpanChunk asyncEvent1_1_1 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 0, 0), new DefaultLocalAsyncId(1, (short) 1));
+        SpanChunk asyncEvent1_1_2 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 0, 1), new DefaultLocalAsyncId(1, (short) 1));
+        SpanChunk asyncEvent1_2_1 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 0, 0), new DefaultLocalAsyncId(1, (short) 2));
 
-        SpanChunk event2 = wrapSpanChunk(traceRoot2, createSpanEvent(traceRoot2, 0, (short) 1));
-        SpanChunk asyncEvent2_1 = wrapSpanChunk(traceRoot2, createAsyncSpanEvent(traceRoot2, 0, (short) 0), new DefaultLocalAsyncId(2, (short) 1));
-        SpanChunk asyncEvent2_2 = wrapSpanChunk(traceRoot2, createAsyncSpanEvent(traceRoot2, 0, (short) 0), new DefaultLocalAsyncId(2, (short) 2));
+        SpanChunk event2 = wrapSpanChunk(traceRoot2, createSpanEvent(traceRoot2, 0, 1));
+        SpanChunk asyncEvent2_1 = wrapSpanChunk(traceRoot2, createAsyncSpanEvent(traceRoot2, 0, 0), new DefaultLocalAsyncId(2, (short) 1));
+        SpanChunk asyncEvent2_2 = wrapSpanChunk(traceRoot2, createAsyncSpanEvent(traceRoot2, 0, 0), new DefaultLocalAsyncId(2, (short) 2));
         final List<SpanType> expectedOrder = Arrays.asList(
                 span,
                 event1,
@@ -174,15 +174,15 @@ public class OrderedSpanRecorderTest {
         final TraceRoot traceRoot2 = new DefaultTraceRoot(traceId2, agentId, startTime2, 0);
 
         Span span1 = createSpan(traceRoot1, startTime1);
-        SpanChunk event1_0 = wrapSpanChunk(traceRoot1, createSpanEvent(traceRoot1, 1, (short) 0));
-        SpanChunk event1_1 = wrapSpanChunk(traceRoot1, createSpanEvent(traceRoot1,2, (short) 1));
-        SpanChunk asyncEvent1_0 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 1, (short) 0), new DefaultLocalAsyncId(1, (short) 1));
-        SpanChunk asyncEvent1_1 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 2, (short) 1), new DefaultLocalAsyncId(1, (short) 1));
+        SpanChunk event1_0 = wrapSpanChunk(traceRoot1, createSpanEvent(traceRoot1, 1, 0));
+        SpanChunk event1_1 = wrapSpanChunk(traceRoot1, createSpanEvent(traceRoot1,2, 1));
+        SpanChunk asyncEvent1_0 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 1, 0), new DefaultLocalAsyncId(1, (short) 1));
+        SpanChunk asyncEvent1_1 = wrapSpanChunk(traceRoot1, createAsyncSpanEvent(traceRoot1, 2, 1), new DefaultLocalAsyncId(1, (short) 1));
 
         Span span2 = createSpan(traceRoot2, startTime2);
-        SpanChunk event2_0 = wrapSpanChunk(traceRoot2, createSpanEvent(traceRoot2, 0, (short) 0));
-        SpanChunk event2_1 = wrapSpanChunk(traceRoot2, createSpanEvent(traceRoot2, 1, (short) 1));
-        SpanChunk asyncEvent2_0 = wrapSpanChunk(traceRoot2, createAsyncSpanEvent(traceRoot2, 0, (short) 0), new DefaultLocalAsyncId(2, (short) 1));
+        SpanChunk event2_0 = wrapSpanChunk(traceRoot2, createSpanEvent(traceRoot2, 0, 0));
+        SpanChunk event2_1 = wrapSpanChunk(traceRoot2, createSpanEvent(traceRoot2, 1, 1));
+        SpanChunk asyncEvent2_0 = wrapSpanChunk(traceRoot2, createAsyncSpanEvent(traceRoot2, 0, 0), new DefaultLocalAsyncId(2, (short) 1));
         final List<SpanType> expectedOrder = Arrays.asList(
                 span1,
                 event1_0,
@@ -213,11 +213,11 @@ public class OrderedSpanRecorderTest {
         assertNull(this.recorder.pop());
     }
 
-    private SpanEvent createSpanEvent(TraceRoot traceRoot1, int startElapsed, short sequence) {
+    private SpanEvent createSpanEvent(TraceRoot traceRoot1, int startElapsed, int sequence) {
         return createAsyncSpanEvent(traceRoot1, startElapsed, sequence);
     }
-    
-    private SpanEvent createAsyncSpanEvent(TraceRoot traceRoot, int startElapsed, short sequence) {
+
+    private SpanEvent createAsyncSpanEvent(TraceRoot traceRoot, int startElapsed, int sequence) {
         Objects.requireNonNull(traceRoot, "traceRoot");
         if (startElapsed < 0) {
             throw new IllegalArgumentException("startElapsed cannot be less than 0");
